requirements.txt
fastapi==0.109.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
pandas==2.1.0
textblob==0.17.1
uvicorn==0.24.0
python-dotenv==1.0.0
bcrypt==4.0.1
